wip checkpoint on tuning
