+++ blackjax/mcmc/mclmc.py
@@ -19,7 +19,7 @@ import jax
 from blackjax.base import SamplingAlgorithm
 from blackjax.mcmc.integrators import IntegratorState, noneuclidean_mclachlan
 from blackjax.types import Array, ArrayLike, PRNGKey
-from blackjax.util import generate_unit_vector, partially_refresh_momentum
+from blackjax.util import full_refresh, generate_unit_vector, partially_refresh_momentum
 
 __all__ = ["MCLMCInfo", "init", "build_kernel", "mclmc"]
 
@@ -46,6 +46,13 @@ class MCLMCInfo(NamedTuple):
 
 def init(x_initial: ArrayLike, logdensity_fn, rng_key):
     l, g = jax.value_and_grad(logdensity_fn)(x_initial)
+
+    jax.debug.print("thing blackjax {x}", x=IntegratorState(
+        position=x_initial,
+        momentum=generate_unit_vector(rng_key, x_initial),
+        logdensity=l,
+        logdensity_grad=g,
+    ))
 
     return IntegratorState(
         position=x_initial,
