+++ blackjax/util.py
@@ -1,6 +1,7 @@
 """Utility functions for BlackJax."""
 from functools import partial
 from typing import Union
+import jax
 
 import jax.numpy as jnp
 from jax import jit, lax
@@ -103,6 +104,18 @@ def generate_unit_vector(
     sample = normal(rng_key, shape=p.shape, dtype=p.dtype)
     return unravel_fn(sample / jnp.linalg.norm(sample))
 
+def full_refresh(d):
+  """Generates a random (isotropic) unit vector."""
+  
+  
+  def rng(random_key):
+      key, subkey = jax.random.split(random_key)
+      u = jax.random.normal(jax.random.PRNGKey(0), shape = (d, ))
+      u /= jnp.sqrt(jnp.sum(jnp.square(u)))
+      return u, key
+  
+    
+  return rng
 
 def partially_refresh_momentum(momentum, rng_key, step_size, L):
     """Adds a small noise to momentum and normalizes.
